Update basic_controller.cpp
I used nh because it is a relative nodehandler. This way you will create two stop services that are linked to each node seperately instead of creating the service when starting the first node then overwriting it when running the second node and ending up with the service working with only the second turtle.

The braces this is clear and just a light omission.